static/ical/foulab.ics: fix recurrence of Open Night events
this was a mess of monthly recurrence rules, which sometimes caused
doubled events late in the month. The ical format supports weekly
recurrence, so that should be used instead, as described in
[RFC 5545 § 3.8.5.3][rfc].

The old way had recurrence on the 1st, 2nd, 3rd, 4th, and -1st Tuesdays
and Thursdays, which means that any month that has only 4 Tuesdays or
Thursdays will have a doubled event on the last Tuesday/Thursday.

The new way just has recurrence every week on Tuesday and Thursday

[rfc]: https://datatracker.ietf.org/doc/html/rfc5545#section-3.8.5.3